==> Notebooks/2.1 Preprocessing with Numeric 0 and Categorical Unknown.ipynb <==
**Load packages**

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder

**Load Data**

train = pd.read_csv("../data/raw/train.csv")
test = pd.read_csv("../data/raw/test.csv")

train.head()

test.head()

Check the count of null values in each variable in Train and Test set

null_dict_train = {}
for col in train.columns:
    null_dict_train[col] = train[col].isnull().sum()

null_dict_test = {}
for col in test.columns:
    null_dict_test[col] = test[col].isnull().sum()

print(null_dict_train)

print(null_dict_test)

### Preprocess Data

We replace the categorical NA values by creating a new category named Unknown and replace the numeric NA values with 0

#treating NA values
train['Type_of_Cab'].fillna("Unknown", inplace=True)
test['Type_of_Cab'].fillna("Unknown", inplace=True)

train['Confidence_Life_Style_Index'].fillna("Unknown", inplace=True)
test['Confidence_Life_Style_Index'].fillna("Unknown", inplace=True)

for col in ['Customer_Since_Months', 'Var1', 'Life_Style_Index']:
    train[col].fillna(0, inplace = True)
    train[col] = train[col].astype('int')

for col in ['Customer_Since_Months', 'Var1', 'Life_Style_Index']:
    test[col].fillna(0, inplace = True)
    test[col] = test[col].astype('int')

train['Var1'] = train['Var1'].astype('int')
test['Var1'] = test['Var1'].astype('int')

train['Customer_Since_Months'] = train['Customer_Since_Months'].astype('int')
test['Customer_Since_Months'] = test['Customer_Since_Months'].astype('int')

train['Customer_Since_Months'] = train['Customer_Since_Months'].astype('int')
test['Customer_Since_Months'] = test['Customer_Since_Months'].astype('int')


We label encode the categorical columns using LabelEncoder function

categorical_cols = ['Type_of_Cab', 'Confidence_Life_Style_Index', 'Destination_Type', 'Gender']

for col in categorical_cols:
  print(col)
  le = LabelEncoder()
  trained_le = le.fit(train[col])
  train[col] = trained_le.transform(train[col])
  test[col] = trained_le.transform(test[col])

train.head(10)

train.to_csv('../data/preprocessed/train_process1.csv', index = False)
test.to_csv('../data/preprocessed/test_process1.csv', index = False)
