==> Notebooks/2.3 Preprocessing using KNN Imputer.ipynb <==
**Loaded packages**

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
from sklearn.impute import KNNImputer

**Load Data**

train = pd.read_csv("../data/raw/train.csv")
test = pd.read_csv("../data/raw/test.csv")

train.head()

test.head()

Check the count of null values in each variable in Train and Test set

null_dict_train = {}
for col in train.columns:
    null_dict_train[col] = train[col].isnull().sum()

null_dict_test = {}
for col in test.columns:
    null_dict_test[col] = test[col].isnull().sum()

print(null_dict_train)

print(null_dict_test)

### Preprocess Data

We first label encode the data manually because the KNN imputer needs int/float values

train['Type_of_Cab'] = train['Type_of_Cab'].replace("A",'1')
train['Type_of_Cab'] = train['Type_of_Cab'].replace("B",'2')
train['Type_of_Cab'] = train['Type_of_Cab'].replace("C",'3')
train['Type_of_Cab'] = train['Type_of_Cab'].replace("D",'4')
train['Type_of_Cab'] = train['Type_of_Cab'].replace("E",'5')
train['Type_of_Cab'] = train['Type_of_Cab'].astype(float).astype('Int64')

test['Type_of_Cab'] = test['Type_of_Cab'].replace("A",'1')
test['Type_of_Cab'] = test['Type_of_Cab'].replace("B",'2')
test['Type_of_Cab'] = test['Type_of_Cab'].replace("C",'3')
test['Type_of_Cab'] = test['Type_of_Cab'].replace("D",'4')
test['Type_of_Cab'] = test['Type_of_Cab'].replace("E",'5')
test['Type_of_Cab'] = test['Type_of_Cab'].astype(float).astype('Int64')


train['Confidence_Life_Style_Index'] = train['Confidence_Life_Style_Index'].replace("A",'1')
train['Confidence_Life_Style_Index'] = train['Confidence_Life_Style_Index'].replace("B",'2')
train['Confidence_Life_Style_Index'] = train['Confidence_Life_Style_Index'].replace("C",'3')
train['Confidence_Life_Style_Index'] = train['Confidence_Life_Style_Index'].astype(float).astype('Int64')

test['Confidence_Life_Style_Index'] = test['Confidence_Life_Style_Index'].replace("A",'1')
test['Confidence_Life_Style_Index'] = test['Confidence_Life_Style_Index'].replace("B",'2')
test['Confidence_Life_Style_Index'] = test['Confidence_Life_Style_Index'].replace("C",'3')
test['Confidence_Life_Style_Index'] = test['Confidence_Life_Style_Index'].astype(float).astype('Int64')

dst_type = train['Destination_Type'].unique().tolist()
dest_vals = [0,4,1,2,6,3,5,10,11,7,8,9,12,13]
dest_map = dict(zip(dst_type, dest_vals))

train['Destination_Type'] = train['Destination_Type'].replace(dest_map)
train['Destination_Type'] = train['Destination_Type'].astype(float).astype('Int64')

test['Destination_Type'] = test['Destination_Type'].replace(dest_map)
test['Destination_Type'] = test['Destination_Type'].astype(float).astype('Int64')

le = LabelEncoder()
trained_le = le.fit(train["Gender"])
train["Gender"] = trained_le.transform(train["Gender"])
test["Gender"] = trained_le.transform(test["Gender"])
train.head(10)

X = train.drop(["Surge_Pricing_Type","Trip_ID"], axis=1)
y = train.Surge_Pricing_Type
X_test = test.drop(["Trip_ID"], axis=1)
Trip_ID = test.Trip_ID

Fitting the KNN Imputer

# define imputer
imputer = KNNImputer(n_neighbors = 10, weights='uniform')
fitted_imputer = imputer.fit(X)
X = fitted_imputer.transform(X)
X_test = fitted_imputer.transform(X_test)

X_imputed = pd.DataFrame(X, columns = train.drop(["Surge_Pricing_Type","Trip_ID"], axis=1).columns)
X_test_imputed = pd.DataFrame(X_test, columns = test.drop(["Trip_ID"], axis=1).columns)

X_imputed.insert(loc = 0, column='Trip_ID', value=train['Trip_ID'])
X_imputed.insert(loc = 13, column='Surge_Pricing_Type', value=train['Surge_Pricing_Type'])
X_imputed

X_test_imputed.insert(loc = 0, column='Trip_ID', value=test['Trip_ID'])
X_test_imputed

X_imputed.to_csv('../data/preprocessed/train_process3.csv', index = False)
X_test_imputed.to_csv('../data/preprocessed/test_process3.csv', index = False)
